requirements.txt
google-adk
google-genai
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
python-dateutil
requests